+++ parameters.cpp
@@ -1,7 +1,7 @@
 #include "parameters.h"
 #include <QDir>
 
-int16_t*** Parameters::voxels;
+int32_t*** Parameters::voxels;
 short int Parameters::size;
 int Parameters::points;
 QString Parameters::algorithm;
